backend/requirements.txt
Flask==2.3.3
yfinance==0.2.50
flask-cors==3.0.10
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0
requests[socks]
pyodbc
Flask-Talisman==1.0.0